refactor: reorganize project structure for better maintainability
