+++ frontend/components/features/departments/DepartmentsDirectory.tsx
@@ -1,8 +1,8 @@
-import React, { useState } from 'react';
-import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
-import { Badge } from '../ui/badge';
-import { Button } from '../ui/button';
-import { Input } from '../ui/input';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { 
   Search, 
   GraduationCap, 
@@ -13,6 +13,7 @@ import {
   MapPin,
   ChevronRight
 } from 'lucide-react';
+import { useState } from 'react';
 
 const DEPARTMENTS = [
   {
